Lots of internal upgrades

- Dependencies cleaning/upgrade. Current Python compatibility is 3.10 -> 3.12
- Switch to uv/pyproject for package management
- Installation of Spacy is now optional, use pip install gismo[spacy] to install it
- Switch documentation to Sphinx+Myst
- Minor change in test_dblp.py

+++ gismo/landmarks.py
@@ -138,16 +138,14 @@ class Landmarks(Corpus):
     >>> landmarks.get_landmarks_by_cluster(gismo) # doctest: +NORMALIZE_WHITESPACE +ELLIPSIS
     {'landmark': {'name': 'Gremlins',
                   'content': 'The Gremlins movie features a Mogwai.'},
-                  'focus': 0.999998...,
+                  'focus': np.float64(0.999998...),
                   'children': [{'landmark': {'name': 'Gremlins',
-                                             'content': 'The Gremlins movie features a Mogwai.'},
-                                             'focus': 1.0, 'children': []},
-                               {'landmark': {'name': 'Star Wars',
-                                             'content': 'The Star Wars movies feature Yoda.'},
-                                             'focus': 1.0, 'children': []},
-                               {'landmark': {'name': 'Movies',
-                                             'content': 'Star Wars, Gremlins, and Blade are movies.'},
-                                             'focus': 1.0, 'children': []}]}
+                  'content': 'The Gremlins movie features a Mogwai.'},
+                  'focus': 1.0, 'children': []},
+                  {'landmark': {'name': 'Star Wars', 'content': 'The Star Wars movies feature Yoda.'},
+                  'focus': 1.0, 'children': []},
+                  {'landmark': {'name': 'Movies', 'content': 'Star Wars, Gremlins, and Blade are movies.'},
+                  'focus': 1.0, 'children': []}]}
 
     We can set the `post_cluster` attribute to customize the output. Gismo provides a simple display.
